Added global vars

+++ GenerateAsmCode.h
@@ -4,29 +4,29 @@
 #include "Tree.h"
 #include "Stack.h"
 
-const size_t INITIAL_CAPACITY = 20;
-const size_t REG_NAME_SIZE    = 2;
-const size_t NUMBERS_VARIABLE = 10;
-const size_t NUMBERS_LABEL    = 10;
+const int INITIAL_CAPACITY = 20;
+const int REG_NAME_SIZE    = 2;
+const int NUMBERS_VARIABLE = 10;
+const int NUMBERS_LABEL    = 10;
 
 struct Name
 {
     char str[STR_MAX_SIZE];
-    size_t curOffset;
+    int curOffset;
 };
 
 struct TableLocalNames
 {
     Name localNames[NUMBERS_VARIABLE];
-    size_t curName;
-    size_t curOffset;
+    int curName;
+    int curOffset;
 };
 
 struct TableGlobalNames
 {
     Name globalNames[NUMBERS_VARIABLE];
-    size_t curName;
-    size_t curOffset;
+    int curName;
+    int curOffset;
 };
 
 void GenerateAsmCode(Tree_t *tree);
